Safety Push
After this I'm going to branch out from master I think so we're making a very minimalistic push. I cleaned up a few random things from the code and I began work on using a community API unfortunately it's causing conflicts with Webpack 5 so I'm going to have to decide what to do about that.